# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: CK Empire Builder API
  description: |
    # CK Empire Builder - Advanced Digital Empire Management API
    
    ## Overview
    
    CK Empire Builder is a comprehensive digital empire management platform that provides:
    
    * **Project Management**: Organize and track digital projects
    * **Revenue Tracking**: Monitor income streams and financial metrics
    * **AI Integration**: Advanced AI features for content generation
    * **Ethics Monitoring**: Built-in ethical AI monitoring
    * **Performance Analytics**: Real-time metrics and insights
    * **Cloud Integration**: AWS/GCP integration with auto-scaling
    * **Backup & Recovery**: Automated backup and disaster recovery
    
    ## Features
    
    ### 🔥 Core Features
    - **Project Management**: Create, track, and manage digital projects
    - **Revenue Analytics**: Monitor income streams and financial performance
    - **AI-Powered Content**: Generate content using advanced AI models
    - **Ethics Compliance**: Built-in ethical AI monitoring and compliance
    - **Performance Monitoring**: Real-time analytics and insights
    
    ### ☁️ Cloud Features
    - **AWS Integration**: S3 backup, RDS database, ElastiCache Redis
    - **GCP Integration**: Cloud Storage, Cloud SQL, Memorystore
    - **Auto-scaling**: Kubernetes HPA for automatic scaling
    - **Backup & Recovery**: Automated database backups and restoration
    - **Monitoring**: CloudWatch metrics and alerts
    
...
servers:
...
security:
  - BearerAuth: []
paths:
...
  /api/v1/projects:
    get:
      summary: List Projects
      description: Retrieve a list of all projects
      tags:
        - Projects
      parameters:
        - name: page
          in: query
          description: Page number for pagination
          required: false
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          description: Number of items per page
          required: false
          schema:
            type: integer
            default: 10
            maximum: 100
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectListResponse'
    
    post:
      summary: Create Project
      description: Create a new project
      tags:
        - Projects
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectCreateRequest'
      responses:
        '201':
          description: Project created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectResponse'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/projects/{project_id}:
    get:
      summary: Get Project
      description: Retrieve a specific project by ID
      tags:
        - Projects
      parameters:
        - name: project_id
          in: path
          required: true
          description: Project ID
          schema:
            type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectResponse'
        '404':
          description: Project not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    
    put:
      summary: Update Project
      description: Update an existing project
      tags:
        - Projects
      parameters:
        - name: project_id
          in: path
          required: true
          description: Project ID
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectUpdateRequest'
      responses:
        '200':
          description: Project updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectResponse'
        '404':
          description: Project not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    
    delete:
      summary: Delete Project
      description: Delete a project
      tags:
        - Projects
      parameters:
        - name: project_id
          in: path
          required: true
          description: Project ID
          schema:
            type: string
      responses:
        '204':
          description: Project deleted successfully
        '404':
          description: Project not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/v1/ai/ideas:
    post:
      summary: Generate AI Ideas
      description: Generate AI-powered ideas for content or projects
      tags:
        - AI
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IdeaGenerationRequest'
      responses:
        '200':
          description: Ideas generated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IdeaGenerationResponse'
  /api/v1/ai/content:
    post:
      summary: Generate AI Content
      description: Generate AI-powered content
      tags:
        - AI
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ContentGenerationRequest'
      responses:
        '200':
          description: Content generated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContentGenerationResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ProjectCreateRequest:
      type: object
      required:
        - name
        - description
      properties:
        name:
          type: string
          description: Project name
          example: "AI Content Generator"
        description:
          type: string
          description: Project description
          example: "Advanced AI-powered content generation system"
        status:
          type: string
          enum: [active, inactive, completed, paused]
          default: active
          example: "active"
        metadata_encrypted:
          type: string
          description: Encrypted metadata
          example: "encrypted-metadata-string"
    ProjectUpdateRequest:
      type: object
      properties:
        name:
          type: string
          description: Project name
          example: "AI Content Generator v2"
        description:
          type: string
          description: Project description
          example: "Enhanced AI-powered content generation system"
        status:
          type: string
          enum: [active, inactive, completed, paused]
          example: "active"
        metadata_encrypted:
          type: string
          description: Encrypted metadata
          example: "encrypted-metadata-string"
    ProjectResponse:
      type: object
      properties:
        id:
          type: string
          example: "proj_123456789"
        name:
          type: string
          example: "AI Content Generator"
        description:
          type: string
          example: "Advanced AI-powered content generation system"
        status:
          type: string
          enum: [active, inactive, completed, paused]
          example: "active"
        created_at:
          type: string
          format: date-time
          example: "2024-01-01T00:00:00Z"
        updated_at:
          type: string
          format: date-time
          example: "2024-01-01T00:00:00Z"
        metadata_encrypted:
          type: string
          example: "encrypted-metadata-string"
    ProjectListResponse:
      type: object
      properties:
        projects:
          type: array
          items:
            $ref: '#/components/schemas/ProjectResponse'
        total:
          type: integer
          example: 25
        page:
          type: integer
          example: 1
        limit:
          type: integer
          example: 10
        pages:
          type: integer
          example: 3
...
    IdeaGenerationRequest:
      type: object
      required:
        - topic
      properties:
        topic:
          type: string
          description: Topic for idea generation
          example: "artificial intelligence"
        style:
          type: string
          enum: [viral, professional, creative, technical]
          default: viral
          example: "viral"
        count:
          type: integer
          minimum: 1
          maximum: 10
          default: 5
          example: 5
    IdeaGenerationResponse:
      type: object
      properties:
        ideas:
          type: array
          items:
            type: object
            properties:
              title:
                type: string
                example: "AI-Powered Personal Assistant"
              description:
                type: string
                example: "Create an AI assistant that learns user preferences"
              category:
                type: string
                example: "Technology"
              potential_revenue:
                type: number
                format: float
                example: 50000.00
        total_generated:
          type: integer
          example: 5
    ContentGenerationRequest:
      type: object
      required:
        - prompt
      properties:
        prompt:
          type: string
          description: Content generation prompt
          example: "Write a blog post about AI trends in 2024"
        content_type:
          type: string
          enum: [blog, social, video, email]
          default: blog
          example: "blog"
        tone:
          type: string
          enum: [professional, casual, formal, creative]
          default: professional
          example: "professional"
        length:
          type: string
          enum: [short, medium, long]
          default: medium
          example: "medium"
    ContentGenerationResponse:
      type: object
      properties:
        content:
          type: string
          example: "Generated content text..."
        title:
          type: string
          example: "AI Trends in 2024: What to Expect"
        summary:
          type: string
          example: "A comprehensive overview of AI trends..."
        word_count:
          type: integer
          example: 1500
        generation_time:
          type: number
          format: float
          example: 2.5
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: "Resource not found"
        message:
          type: string
          example: "The requested resource was not found"
        status_code:
          type: integer
          example: 404
        timestamp:
          type: string
          format: date-time
          example: "2024-01-01T00:00:00Z"
...